Move getContract into utils.js
